+++ CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Add 'duplicate' button
+
 ## v1.0.0
 
 - Initial release